refactor: refactor services and change folder structure
