+++ chat/src/main/kotlin/net/thechance/chat/api/dto/ChatSummaryResponse.kt
@@ -1,6 +1,6 @@
 package net.thechance.chat.api.dto
 
-import net.thechance.chat.entity.ChatSummary
+import net.thechance.chat.service.model.ChatSummary
 import org.springframework.data.domain.Page
 import java.time.Instant
 import java.util.*
