+++ chat/src/main/kotlin/net/thechance/chat/service/ContactService.kt
@@ -19,7 +19,7 @@ class ContactService(
     fun getPagedContactByUserId(userId: UUID, pageable: Pageable): Page<ContactModel> {
         return contactRepository.findAllContactModelsByContactOwnerId(
             userId,
-            PageRequest.of(pageable.pageNumber, pageable.pageSize, Sort.by("firstName").ascending())
+            PageRequest.of(pageable.pageNumber, pageable.pageSize, Sort.by(Contact::firstName.name).ascending())
         )
     }
