+++ chat/src/main/kotlin/net/thechance/chat/service/model/ChatSummary.kt
@@ -1,11 +1,24 @@
-package net.thechance.chat.service
+package net.thechance.chat.service.model
 
 import net.thechance.chat.entity.Chat
-import net.thechance.chat.entity.ChatSummary
 import net.thechance.chat.entity.ContactUser
 import net.thechance.chat.entity.Message
-import java.util.*
+import java.time.Instant
+import java.util.UUID
 
+data class ChatSummary(
+    val id: UUID,
+    val name: String,
+    val imageUrl: String?,
+    val lastMessage: Message?,
+    val unReadMessagesCount: Int
+) {
+    data class Message(
+        val text: String,
+        val sentAt: Instant,
+        val isMine: Boolean
+    )
+}
 fun Chat.toSummary(userId: UUID, otherUser: ContactUser?, lastMessage: Message?, unreadCount: Int): ChatSummary {
     return ChatSummary(
         id = id,
@@ -21,3 +34,5 @@ fun Chat.toSummary(userId: UUID, otherUser: ContactUser?, lastMessage: Message?,
         unReadMessagesCount = unreadCount
     )
 }
+
+
